+++ jsonschema-generator/src/test/java/com/github/victools/jsonschema/generator/FieldScopeTest.java
@@ -51,7 +51,11 @@ public class FieldScopeTest extends AbstractTypeAwareTest {
             Arguments.of("fieldWithPublicGetter", "fieldWithoutGetter", "getFieldWithPublicGetter"),
             Arguments.of("fieldWithPublicBooleanGetter", null, "isFieldWithPublicBooleanGetter"),
             Arguments.of("isFieldWithMatchingGetter", null, "isFieldWithMatchingGetter"),
-            Arguments.of("isFieldWithMatchingGetter", "fieldWithPublicGetter", "isFieldWithMatchingGetter")
+            Arguments.of("isFieldWithMatchingGetter", "fieldWithPublicGetter", "isFieldWithMatchingGetter"),
+            Arguments.of("aFieldWithJavaBeansConformingGetter", null, "getaFieldWithJavaBeansConformingGetter"),
+            Arguments.of("aFieldWithJavaBeansNonConformingGetter", null, "getAFieldWithJavaBeansNonConformingGetter"),
+            Arguments.of("aFieldWithJavaBeansConformingBooleanGetter", null, "isaFieldWithJavaBeansConformingBooleanGetter"),
+            Arguments.of("aFieldWithJavaBeansNonConformingBooleanGetter", null, "isAFieldWithJavaBeansNonConformingBooleanGetter")
         );
     }
 
@@ -110,6 +114,10 @@ public class FieldScopeTest extends AbstractTypeAwareTest {
         private long fieldWithPublicGetter;
         private boolean fieldWithPublicBooleanGetter;
         private boolean isFieldWithMatchingGetter;
+        private int aFieldWithJavaBeansConformingGetter;
+        private long aFieldWithJavaBeansNonConformingGetter;
+        private boolean aFieldWithJavaBeansConformingBooleanGetter;
+        private boolean aFieldWithJavaBeansNonConformingBooleanGetter;
 
         private int getFieldWithPrivateGetter() {
             return this.fieldWithPrivateGetter;
@@ -126,6 +134,22 @@ public class FieldScopeTest extends AbstractTypeAwareTest {
 
         public boolean isFieldWithMatchingGetter() {
             return this.isFieldWithMatchingGetter;
+        }
+
+        public int getaFieldWithJavaBeansConformingGetter() {
+            return this.aFieldWithJavaBeansConformingGetter;
+        }
+
+        public long getAFieldWithJavaBeansNonConformingGetter() {
+            return this.aFieldWithJavaBeansNonConformingGetter;
+        }
+
+        public boolean isaFieldWithJavaBeansConformingBooleanGetter() {
+            return this.aFieldWithJavaBeansConformingBooleanGetter;
+        }
+
+        public boolean isAFieldWithJavaBeansNonConformingBooleanGetter() {
+            return this.aFieldWithJavaBeansNonConformingBooleanGetter;
         }
     }