+++ jsonschema-generator/src/test/java/com/github/victools/jsonschema/generator/MethodScopeTest.java
@@ -52,6 +52,10 @@ public class MethodScopeTest extends AbstractTypeAwareTest {
             Arguments.of("isFieldWithPublicBooleanGetter", "isBehavingSomehow", "fieldWithPublicBooleanGetter"),
             Arguments.of("isFieldWithMatchingPublicGetter", null, "isFieldWithMatchingPublicGetter"),
             Arguments.of("isFieldWithMatchingPublicGetter", "isFieldWithPublicBooleanGetter", "isFieldWithMatchingPublicGetter"),
+            Arguments.of("getaFieldWithJavaBeansConformingGetter", null, "aFieldWithJavaBeansConformingGetter"),
+            Arguments.of("getAFieldWithJavaBeansNonConformingGetter", null, "aFieldWithJavaBeansNonConformingGetter"),
+            Arguments.of("isaFieldWithJavaBeansConformingBooleanGetter", null, "aFieldWithJavaBeansConformingBooleanGetter"),
+            Arguments.of("isAFieldWithJavaBeansNonConformingBooleanGetter", null, "aFieldWithJavaBeansNonConformingBooleanGetter"),
             Arguments.of("getCalculatedValue", null, null),
             Arguments.of("isBehavingSomehow", null, null),
             Arguments.of("isBehavingSomehow", "isFieldWithPublicBooleanGetter", null),
@@ -119,6 +123,10 @@ public class MethodScopeTest extends AbstractTypeAwareTest {
         @TestAnnotation
         private boolean fieldWithPublicBooleanGetter;
         private boolean isFieldWithMatchingPublicGetter;
+        private int aFieldWithJavaBeansConformingGetter;
+        private long aFieldWithJavaBeansNonConformingGetter;
+        private boolean aFieldWithJavaBeansConformingBooleanGetter;
+        private boolean aFieldWithJavaBeansNonConformingBooleanGetter;
 
         @TestAnnotation
         private int getFieldWithPrivateGetter() {
@@ -135,6 +143,22 @@ public class MethodScopeTest extends AbstractTypeAwareTest {
 
         public boolean isFieldWithMatchingPublicGetter() {
             return this.isFieldWithMatchingPublicGetter;
+        }
+
+        public int getaFieldWithJavaBeansConformingGetter() {
+            return this.aFieldWithJavaBeansConformingGetter;
+        }
+
+        public long getAFieldWithJavaBeansNonConformingGetter() {
+            return this.aFieldWithJavaBeansNonConformingGetter;
+        }
+
+        public boolean isaFieldWithJavaBeansConformingBooleanGetter() {
+            return this.aFieldWithJavaBeansConformingBooleanGetter;
+        }
+
+        public boolean isAFieldWithJavaBeansNonConformingBooleanGetter() {
+            return this.aFieldWithJavaBeansNonConformingBooleanGetter;
         }
 
         public double getCalculatedValue() {